Rename ReadLink to ReadAndFollowLink and add Readlink
This was done because the ReadLink function behaved
differently than os.Readlink (lowercase l). The os.Readlink
function does not follow links, whereas the ReadLink method did.
This was confusing, so the ReadLink method was renamed and the
Readlink was added, which matches the behavior of the os.Readlink
function.

[#127235101]

Signed-off-by: Tyler Schultz <tschultz@pivotal.io>