Add volume name to symlinked paths on Windows
Windows creates the wrong type of symlink if the paths are not absolute.
On Windows "/var/vcap", "\var\vcap" and "C:\var\vcap" are considered the
same path (assuming the current volume is C:).  For whatever reason when
creating symlinks on Windows, the paths must be absolute and include
the volume letter - otherwise Windows will create file symlinks for
directories.

Signed-off-by: Charlie Vieth <cviethjr@pivotal.io>