Record counts of files written and commands run.
- It would be an onerous amount of work to pass a logger into the fakes,
and perhaps a failure to separate concerns. But we need a way to know
whether we wrote out files quietly, and ran commands quietly. So we
provide means to record the commands that were run (whether quietly or
not), and likewise count the number of times the quiet and non-quiet
versions of WriteFile were run.

[#149129083](https://www.pivotaltracker.com/story/show/149129083)

Signed-off-by: Arthur Chan <achan@pivotal.io>