Update fake ReadLink to return error similar to real implementation
When ReadLink fails to find the location of the symlink, the fake
implementation returned a custom error. However, the real one
returned an error that matched `os.ErrNotExist`. As such, the fake
implementation was aligned to the real one.

This update also makes ReadAndFollowLink and ReadLink return the same
error when the code fails to find the link on the target system.

[#162161084]