🐞 Honor `NO_PROXY`
In spite of our
[documentation](https://bosh.io/docs/cli-global-flags/#http-proxy)
which said that we honored `NO_PROXY`, we didn't. This commit fixes
that.

We now honor `NO_PROXY`, and if it's not set, we fallback to honoring
`no_proxy`. This brings the behavior in line with our documentation as
well as
[Golang's](https://pkg.go.dev/golang.org/x/net/http/httpproxy#Config).

fixes [#78] Should bosh-utils respect $NO_PROXY instead of $no_proxy?

[#177737289](https://www.pivotaltracker.com/story/show/177737289)

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>