🐞 Prevent test pollution by resetting DialContext
On 9/8, we introduced a new feature which allowed the DialContext to be
reloaded. Unfortunately, the test for the reload left the Dialer in a
modified state. This could cause subsequent tests to intermittently
fail.

This commit fixes that error by unsetting the environment variable
(`BOSH_ALL_PROXY`) and then calling `httpclient.ResetDialerContext()` to leave the
DialContext in a pristine state.

Fixes:
```
Linux-specific tests [It] enables TCP (socket) keepalive with an appropriate interval
/tmp/build/1944c6a2/gopath/src/github.com/cloudfoundry/bosh-utils/httpclient/keepalive_syscall_linux_test.go:18

  Unexpected error:
  ...
      socks connect tcp 127.0.0.1:22->127.0.0.1:19642: dial tcp 127.0.0.1:22: connect: connection refused
  occurred
```

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>